refactor(organizations): extract mapOrganizationToState and align test file naming
Address review feedback on #152:
- Extract a testable mapOrganizationToState helper (symmetry with the
  workspaces mapWorkspaceToState pattern); the unit test now exercises it
  directly instead of stopping at the API-response unmarshal.
- Rename the test files to the repo-consistent convention: the internal
  unit test becomes organization_data_source_mapping_test.go (mirroring
  agent_resource_mapping_test.go) and the external acceptance test takes
  the plain organization_data_source_test.go suffix.
